Fixing the displaying images in the question card and question page when there are more than 2 images
