--- app/api/deleteImage.ts
@@ -1,40 +1,0 @@
-import type { NextApiRequest, NextApiResponse } from "next";
-import cloudinary from "cloudinary";
-
-cloudinary.v2.config({
-  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
-  api_key: process.env.CLOUDINARY_API_KEY,
-  api_secret: process.env.CLOUDINARY_API_SECRET,
-});
-
-type Data = {
-  message?: string;
-  error?: string;
-};
-
-export default async function handler(
-  req: NextApiRequest,
-  res: NextApiResponse<Data>
-) {
-  if (req.method === "POST") {
-    try {
-      const requestBody = JSON.parse(req.body);
-
-      if (!requestBody || !requestBody.publicId) {
-        return res.status(400).json({ error: "Missing public ID" });
-      }
-
-      const publicId: string = requestBody.publicId;
-
-      await cloudinary.v2.uploader.destroy(publicId);
-
-      res.status(200).json({ message: "Image deleted successfully" });
-    } catch (error) {
-      console.error("Error deleting image:", error);
-      res.status(500).json({ error: "Error deleting image" });
-    }
-  } else {
-    res.setHeader("Allow", ["POST"]);
-    res.status(405).end(`Method ${req.method} Not Allowed`);
-  }
-}
