+++ components/community/SearchQuestion.tsx
@@ -37,40 +37,6 @@ const SearchQuestion = ({ question }: SearchQuestionProps) => {
   const timeDifference = useFormatedDate(question.createdAt);
   const { data: session } = useSession();
 
-  const handleDeleteFromFirebase =
-    (itemId: string, image?: string) => async () => {
-      if (session?.user?.id) {
-        try {
-          const questionRef = doc(db, "questions", itemId);
-          await deleteDoc(questionRef);
-        } catch (error) {
-          console.error("Error deleting post:", error);
-        }
-        //Work on delete image function here
-        // const match = image?.match(/circhoolar_items_upload\/(.+)\.jpg/);
-        // if (match && match.length >= 2) {
-        //   const publicId = match[1];
-        //   try {
-        //     // Send DELETE request to Cloudinary
-        //     await fetch("/api/deleteImage", {
-        //       method: "POST",
-        //       headers: {
-        //         "Content-Type": "application/json",
-        //       },
-        //       body: JSON.stringify({ publicId }),
-        //     });
-        //     // Delete the post from Firebase
-        //     const questionRef = doc(db, "posts", itemId);
-        //     await deleteDoc(questionRef);
-        //   } catch (error) {
-        //     console.error("Error deleting post:", error);
-        //   }
-        // } else {
-        //   console.error("Invalid image URL format");
-        // }
-      }
-    };
-
   return (
     <Card className="bg-light-white border border-gray-50 shadow-sm hover:shadow-md hover:border-paragraph-color rounded-md p-4 mt-2">
       {session?.user?.name === question.author ? (
